+++ xbmc/cores/VideoPlayer/DVDDemuxers/DVDDemuxClient.h
@@ -37,7 +37,7 @@ public:
   void FillBuffer(bool mode) override;
   CDemuxStream* GetStream(int iStreamId) const override;
   std::vector<CDemuxStream*> GetStreams() const override;
-  int GetNrOfStreams() const override;
+  size_t GetNrOfStreams() const override;
   std::string GetFileName() override;
   std::string GetStreamCodecName(int iStreamId) override;
   void EnableStream(int id, bool enable) override;
